Initial Deliverable 3 commit
