+++ Service_Novigrad/app/src/main/java/com/example/servicenovigrad/ui/RegisterCustomer.java
@@ -14,7 +14,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.servicenovigrad.R;
 import com.example.servicenovigrad.users.Customer;
-import com.example.servicenovigrad.destinations.CustomerHomePage;
+import com.example.servicenovigrad.ui.homepages.CustomerHomePage;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
