+++ Service_Novigrad/app/src/main/java/com/example/servicenovigrad/ui/admin/AdminEditBranchAccounts.java
@@ -1,8 +1,9 @@
-package com.example.servicenovigrad.ui;
+package com.example.servicenovigrad.ui.admin;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import com.example.servicenovigrad.R;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -15,51 +16,50 @@ import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.servicenovigrad.R;
-import com.example.servicenovigrad.users.Customer;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+
+import com.example.servicenovigrad.users.BranchEmployee;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
+public class AdminEditBranchAccounts extends AppCompatActivity {
 
-public class AdminEditCustomerAccounts extends AppCompatActivity {
-
-    ListView admin_customer_list;
-    Button admin_delete_customer_btn, admin_customer_cancel_btn,
+    ListView admin_branch_list;
+    Button admin_delete_branch_btn, admin_edit_cancel_btn,
                 confirm_yes_btn, confirm_no_btn;
-    TextView edit_customer_customerName,
-                edit_customer_userName, are_you_sure;
+    TextView edit_branch_branchName, edit_branch_employeeName,
+                edit_branch_userName, are_you_sure;
     Dialog dialog, confirmDelete;
     DatabaseReference allUsersReference;
-    ArrayList<Customer> allCustomers;
-    ArrayAdapter<Customer> arrayAdapter;
+    ArrayList<BranchEmployee> allBranchEmployees;
+    ArrayAdapter<BranchEmployee> arrayAdapter;
 
     //This HashMap lets you get the database ID of a user
-    HashMap<Customer, String> customerIDs;
+    HashMap<BranchEmployee, String> employeeIDs;
 
-    //Used when dealing with a specific customer
-    Customer currCustomer;
-    String currCustomerID;
+    //Used when dealing with a specific employee
+    BranchEmployee currBranchEmployee;
+    String currEmployeeID;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_admin_edit_customer_accounts);
+        setContentView(R.layout.activity_admin_edit_branch_accounts);
+
         allUsersReference = FirebaseDatabase.getInstance().getReference().child("users");
-        admin_customer_list = findViewById(R.id.admin_customer_list);
+        admin_branch_list = findViewById(R.id.admin_branch_list);
 
         allUsersReference.addValueEventListener(new ValueEventListener() {
             @Override
             @SuppressWarnings("unchecked")
             public void onDataChange(@NonNull DataSnapshot snapshot) {
-                //Clear local data after a deletion
-                allCustomers = new ArrayList<>();
-                customerIDs = new HashMap<>();
+                allBranchEmployees = new ArrayList<>();
+                employeeIDs = new HashMap<>();
 
                 //Get JSON tree of all users (Firebase sends it as nested HashMaps)
                 HashMap<String,Object> allUsers = (HashMap<String,Object>) snapshot.getValue();
@@ -68,33 +68,33 @@ public class AdminEditCustomerAccounts extends AppCompatActivity {
                 for (Map.Entry<String, Object> user : allUsers.entrySet()){
                     //Convert each value to a HashMap of strings (this is the user object)
                     HashMap<String, String> userData = (HashMap<String,String>) user.getValue();
-                    //Find the customers and add them to both collections
-                    if (userData.get("role").equals("CUSTOMER")){
-                        Customer customer = new Customer(
-                                userData.get("firstName"), userData.get("lastName"),
-                                userData.get("userName"));
-                        allCustomers.add(customer);
-                        customerIDs.put(customer, user.getKey());
+                    //Find the branch employees and add them to both collections
+                    if (userData.get("role").equals("BRANCH_EMPLOYEE")){
+                        BranchEmployee branchEmployee = new BranchEmployee(
+                            userData.get("firstName"), userData.get("lastName"),
+                            userData.get("userName"), userData.get("branchName"));
+                        allBranchEmployees.add(branchEmployee);
+                        employeeIDs.put(branchEmployee, user.getKey());
                     }
                 }
                 //Put the data into the list view
-                arrayAdapter = new ArrayAdapter<>(AdminEditCustomerAccounts.this,
-                        android.R.layout.simple_expandable_list_item_1, allCustomers);
-                admin_customer_list.setAdapter(arrayAdapter);
+                arrayAdapter = new ArrayAdapter<>(AdminEditBranchAccounts.this,
+                    android.R.layout.simple_expandable_list_item_1, allBranchEmployees);
+                admin_branch_list.setAdapter(arrayAdapter);
 
                 //Set on click listener for all items
-                admin_customer_list.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+                admin_branch_list.setOnItemClickListener(new AdapterView.OnItemClickListener() {
                     @Override
                     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                        currCustomer = (Customer) admin_customer_list.getItemAtPosition(position);
-                        currCustomerID = customerIDs.get(currCustomer);
+                        currBranchEmployee = (BranchEmployee) admin_branch_list.getItemAtPosition(position);
+                        currEmployeeID = employeeIDs.get(currBranchEmployee);
                         openDialog();
                     }
                 });
             }
             @Override
             public void onCancelled(@NonNull DatabaseError error) {
-                Toast.makeText(AdminEditCustomerAccounts.this, "Error retrieving data...", Toast.LENGTH_SHORT).show();
+                Toast.makeText(AdminEditBranchAccounts.this, "Error retrieving data...", Toast.LENGTH_SHORT).show();
             }
         });
     }
@@ -104,28 +104,30 @@ public class AdminEditCustomerAccounts extends AppCompatActivity {
         if (dialog != null) dialog.dismiss();
         //Display dialog
         dialog = new Dialog(this);
-        dialog.setContentView(R.layout.dialog_admin_edit_customer);
+        dialog.setContentView(R.layout.dialog_admin_edit_branch);
 
         //Get views
-        edit_customer_customerName = dialog.findViewById(R.id.edit_customer_customerName);
-        edit_customer_userName = dialog.findViewById(R.id.edit_customer_userName);
-        admin_delete_customer_btn = dialog.findViewById(R.id.admin_delete_customer_btn);
-        admin_customer_cancel_btn = dialog.findViewById(R.id.admin_customer_cancel_btn);
+        edit_branch_branchName = dialog.findViewById(R.id.edit_branch_branchName);
+        edit_branch_employeeName = dialog.findViewById(R.id.edit_branch_employeeName);
+        edit_branch_userName = dialog.findViewById(R.id.edit_branch_userName);
+        admin_delete_branch_btn = dialog.findViewById(R.id.admin_delete_branch_btn);
+        admin_edit_cancel_btn = dialog.findViewById(R.id.admin_edit_cancel_btn);
 
         //Set text
-        edit_customer_customerName.setText(currCustomer.getFullName());
-        edit_customer_userName.setText(currCustomer.getUserName());
+        edit_branch_branchName.setText(currBranchEmployee.getBranchName());
+        edit_branch_employeeName.setText(currBranchEmployee.getFullName());
+        edit_branch_userName.setText(currBranchEmployee.getUserName());
 
         //Listeners
         //Delete button
-        admin_delete_customer_btn.setOnClickListener(new View.OnClickListener() {
+        admin_delete_branch_btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 openConfirmation();
             }
         });
         //Cancel button
-        admin_customer_cancel_btn.setOnClickListener(new View.OnClickListener() {
+        admin_edit_cancel_btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 dialog.dismiss();
@@ -145,7 +147,7 @@ public class AdminEditCustomerAccounts extends AppCompatActivity {
         confirm_yes_btn = confirmDelete.findViewById(R.id.confirm_yes_btn);
         confirm_no_btn = confirmDelete.findViewById(R.id.confirm_no_btn);
         //Set text
-        are_you_sure.setText(getString(R.string.are_you_sure_customer, currCustomer.getFullName()));
+        are_you_sure.setText(getString(R.string.are_you_sure_branch, currBranchEmployee.getBranchName()));
 
         //Listeners
         //yes button
@@ -170,24 +172,27 @@ public class AdminEditCustomerAccounts extends AppCompatActivity {
     public void deleteUser(){
         //Firebase calls OnDataChange when a value is deleted
         //So the ListView will auto update
-        allUsersReference.child(currCustomerID).removeValue(new DatabaseReference.CompletionListener() {
+        allUsersReference.child(currEmployeeID).removeValue(new DatabaseReference.CompletionListener() {
             @Override
             public void onComplete(@Nullable DatabaseError error, @NonNull DatabaseReference ref) {
                 if (error == null){
                     confirmDelete.dismiss();
                     dialog.dismiss();
-                    Toast.makeText(AdminEditCustomerAccounts.this,
-                            getString(R.string.customer_deleted, currCustomer.getFullName()),
-                            Toast.LENGTH_SHORT).show();
-                    currCustomer = null;
-                    currCustomerID = null;
+                    Toast.makeText(AdminEditBranchAccounts.this,
+                        getString(R.string.branch_deleted, currBranchEmployee.getBranchName()),
+                        Toast.LENGTH_SHORT).show();
+                    currBranchEmployee = null;
+                    currEmployeeID = null;
                 }
                 else{
-                    Toast.makeText(AdminEditCustomerAccounts.this, "Error deleting user...", Toast.LENGTH_SHORT).show();
                     Log.e("TAG", error.toString());
+                    Toast.makeText(AdminEditBranchAccounts.this,
+                        "Error deleting user...", Toast.LENGTH_SHORT).show();
                 }
+
             }
         });
+
     }
 
     //Close any hidden dialogs
