+++ Service_Novigrad/app/src/main/java/com/example/servicenovigrad/ui/branchEmployee/ServicesOffered.java
@@ -1,0 +1,6 @@
+package com.example.servicenovigrad.ui.branchEmployee;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class ServicesOffered extends AppCompatActivity {
+}
