+++ Service_Novigrad/app/src/main/java/com/example/servicenovigrad/users/BranchEmployee.java
@@ -2,7 +2,7 @@
  * BranchEmployee class is for employee users and a subclass of Account class
  * It stores employee credential information. It can create branch account and use
  * methods of the Branch class to edit it.
- */
+ **/
 
 package com.example.servicenovigrad.users;
 
@@ -10,7 +10,25 @@ public class BranchEmployee extends Account {
 
     private final static Roles role = Roles.BRANCH_EMPLOYEE;
 
-    private String branchName;
+    private String branchName, address, phoneNumber;
+
+    /**
+     * Constructor
+     *
+     * @param firstName
+     * @param lastName
+     * @param userName
+     * @param branchName
+     * @param address
+     * @param phoneNumber
+     */
+
+    public BranchEmployee(String firstName, String lastName, String userName, String branchName, String address, String phoneNumber) {
+        super(firstName, lastName, userName, role);
+        this.branchName = branchName;
+        this.address = address;
+        this.phoneNumber = phoneNumber;
+    }
 
     /**
      * Constructor
@@ -21,14 +39,17 @@ public class BranchEmployee extends Account {
      * @param branchName
      */
     public BranchEmployee(String firstName, String lastName, String userName, String branchName) {
-        super(firstName, lastName, userName, role);
-        this.branchName = branchName;
+        this(firstName, lastName, userName, branchName, null, null);
     }
 
 
     public String getBranchName() {
         return branchName;
     }
+
+    public String getAddress() { return address; }
+
+    public String getPhoneNumber() { return phoneNumber; }
 
     @Override
     public String toString(){
