+++ Service_Novigrad/app/src/main/res/layout/activity_admin_edit_branch_accounts.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AdminEditBranchAccounts">
+    tools:context=".ui.admin.AdminEditBranchAccounts">
 
     <LinearLayout
         android:layout_width="match_parent"
