+++ Service_Novigrad/app/src/main/res/layout/activity_admin_edit_customer_accounts.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AdminEditCustomerAccounts">
+    tools:context=".ui.admin.AdminEditCustomerAccounts">
 
     <LinearLayout
         android:layout_width="match_parent"
