+++ Service_Novigrad/app/src/main/res/layout/activity_branch_info.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.admin.AdminEditService">
+    tools:context=".ui.branchEmployee.BranchInfo">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -17,7 +18,7 @@
             android:orientation="horizontal">
 
             <TextView
-                android:id="@+id/admin_edit_service_header"
+                android:id="@+id/branch_name"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
@@ -28,7 +29,7 @@
                 android:textStyle="bold" />
 
             <Button
-                android:id="@+id/admin_del_service"
+                android:id="@+id/edit_branch_name"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="5dp"
@@ -36,7 +37,7 @@
                 android:layout_weight="0.3"
                 android:minWidth="0dp"
                 android:minHeight="0dp"
-                android:text="@string/delete" />
+                android:text="@string/edit" />
 
         </LinearLayout>
 
@@ -47,18 +48,18 @@
             android:orientation="horizontal">
 
             <TextView
-                android:id="@+id/admin_form_field_header"
+                android:id="@+id/branch_address"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:layout_marginLeft="10dp"
                 android:layout_weight="1"
-                android:text="@string/form_fields"
+                android:text="@string/placeholder"
                 android:textSize="18sp"
                 android:textStyle="bold" />
 
             <Button
-                android:id="@+id/admin_add_form_field"
+                android:id="@+id/edit_branch_address"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="5dp"
@@ -66,17 +67,9 @@
                 android:layout_weight="0.3"
                 android:minWidth="0dp"
                 android:minHeight="0dp"
-                android:text="@string/add"
+                android:text="@string/edit"
                 android:textSize="14sp" />
         </LinearLayout>
-
-        <ListView
-            android:id="@+id/admin_form_field_list"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-
-        </ListView>
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -85,18 +78,18 @@
             android:orientation="horizontal">
 
             <TextView
-                android:id="@+id/admin_doc_type_header"
+                android:id="@+id/branch_phone_number"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:layout_marginLeft="10dp"
                 android:layout_weight="1"
-                android:text="@string/document_types"
+                android:text="@string/placeholder"
                 android:textSize="18sp"
                 android:textStyle="bold" />
 
             <Button
-                android:id="@+id/admin_add_doc_type"
+                android:id="@+id/edit_branch_phone_number"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="5dp"
@@ -104,22 +97,10 @@
                 android:layout_weight="0.3"
                 android:minWidth="0dp"
                 android:minHeight="0dp"
-                android:text="@string/add"
+                android:text="@string/edit"
                 android:textSize="14sp" />
         </LinearLayout>
-
-        <ListView
-            android:id="@+id/admin_doc_type_list"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1">
-
-        </ListView>
 
     </LinearLayout>
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal"></LinearLayout>
 </RelativeLayout>
